requirements.txt
numpy
Pillow
scipy
shapely
rasterio
geopandas
pyogrio

